Upload 1789.py

+++ BOJ/Python/Greedy/1789.py
@@ -1,7 +1,8 @@
 import math
 
+
 def solution(S: int) -> int:
-    return (-1 + math.sqrt(1+4*2*S)) // 2
+    return int((-1 + math.sqrt(1+4*2*S)) // 2)
 
 
 if __name__ == '__main__':
